Move some stuff to other files
Move player related stuff to `player.cpp`/`player.hpp`. Because of
this I also needed to create `misc.hpp` to keep stuff all files need.
